pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Andrew2797 <andrijgolub27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-restful = "^0.3.10"
flask-sqlalchemy = "^3.1.1"
requests-html = "^0.10.0"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
lxml = {extras = ["html-clean"], version = "^5.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
